Implemented simple escaping in strings.
A backslash in a string now makes it so the next character -- whatever it may be -- is put into the string directly.
Right now this is only useful for ", as there's no implementation for \n or \t or so on. But then, an actual
return or tab character are not treated specially either.