Added a 'reference' node type.
The reference node type is for referring to variables. Prefixed with @ or $ and currently allowed to consist of letters, numbers, and underscores. It will later be more selective in what can be in it, most likely. And will probably allow an extra $ or @ at the start to indicate global and class variable references respectively.