Implemented reversable recursive inspect and used it to generate test data.
inspect_r on a parse tree will now output a tree that uses the NodeType[] syntax as description, which can be piped back into ruby to
regenerate a parse tree. I then used this to create parse tree data for the sample files so I can tell if a change causes a major
change in how documents are structured.