Implemented an arbitrary quoted string.
Ruby uses %, but I don't like the ambiguity that introduces with the modulus operator, so I've used # instead.
It's required that one of these quotes have a letter after it, and then a character that indicates a terminator.

Also made it so that escaping ONLY happens on the terminator, passing everything else out to higher levels. This
allows a lot more flexibility for the upper layer, especially when it comes to things like regexes. It's expected
that a higher level will implement regexes with something like %r/regex/ as (again), the bare / syntax leaves
a lot of ambiguities in the grammer.